requirements.txt
# UI
streamlit==1.39.0

# Data
numpy==1.26.4
pandas==2.2.3

# Vector search + embeddings
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
torch>=2.2.0

# LLM client
requests==2.32.3
python-dotenv==1.0.1

# Web ingest / scraping
beautifulsoup4==4.12.3
trafilatura==1.12.2
lxml==5.3.0

# Telegram ingest
telethon==1.36.0

# Plots
matplotlib>=3.8

# PDF ingest (PyMuPDF -> fitz)
pymupdf==1.24.9

# DOCX ingest
python-docx==1.1.2

# faster regex + safe HTML parsing
regex>=2024.7.24

# for some pdf extraction fallback
pypdf>=4.3.1
tqdm>=4.66.5

